Revert "Patch 1"
