+++ api/cmd/helix/serve.go
@@ -228,18 +228,12 @@ func serve(cmd *cobra.Command, cfg *config.ServerConfig) error {
 		return fmt.Errorf("runner token is required")
 	}
 
-	gocloakClient, err := auth.NewGoCloakClient(&cfg.Keycloak)
+	keycloakAuthenticator, err := auth.NewKeycloakAuthenticator(&cfg.Keycloak)
 	if err != nil {
-		return fmt.Errorf("failed to create keycloak client: %v", err)
+		return fmt.Errorf("failed to create keycloak authenticator: %v", err)
 	}
 
-	userRetriever, err := auth.NewKeycloakUserRetriever(gocloakClient.Client, &cfg.Keycloak)
-
-	if err != nil {
-		return fmt.Errorf("failed to create user retriever: %v", err)
-	}
-
-	notifier, err := notification.New(&cfg.Notifications, userRetriever)
+	notifier, err := notification.New(&cfg.Notifications, keycloakAuthenticator)
 	if err != nil {
 		return fmt.Errorf("failed to create notifier: %v", err)
 	}
@@ -431,41 +425,6 @@ func serve(cmd *cobra.Command, cfg *config.ServerConfig) error {
 		defer pingService.Stop()
 	}
 
-	adminConfig := auth.AdminConfig{
-		AdminUserIDs: cfg.WebServer.AdminIDs,
-		AdminUserSrc: cfg.WebServer.AdminSrc,
-	}
-
-	var oidcAuthenticator auth.OIDCAuthenticator
-
-	if cfg.OIDC.Enabled {
-		oidcAuthenticator, err = auth.NewCustomOIDCAuthenticator(&cfg.OIDC, &adminConfig)
-		if err != nil {
-			return fmt.Errorf("unable to create oidc authenticator: %v", err)
-		}
-	} else {
-		oidcAuthenticator, err = auth.NewKeycloakAuthenticator(
-			gocloakClient.Client,
-			&cfg.Keycloak,
-			gocloakClient.Token,
-			userRetriever,
-			&adminConfig,
-		)
-		if err != nil {
-			return fmt.Errorf("failed to create keycloak authenticator: %v", err)
-		}
-	}
-
-	authConfig := &server.AuthConfig {
-		OIDCAuth: oidcAuthenticator,
-		RunnerAuth: auth.RunnerTokenAuthenticator { RunnerToken: cfg.WebServer.RunnerToken },
-		ApiKeyAuth: auth.ApiKeyAuthenticator {
-			Store: store,
-			UserRetriever: userRetriever,
-			AdminConfig: &adminConfig,
-		},
-	}
-
 	server, err := server.NewServer(
 		cfg,
 		postgresStore,
@@ -473,7 +432,7 @@ func serve(cmd *cobra.Command, cfg *config.ServerConfig) error {
 		gse,
 		providerManager,
 		helixInference,
-		authConfig,
+		keycloakAuthenticator,
 		stripe,
 		appController,
 		janitor,
