+++ api/pkg/auth/auth.go
@@ -1,0 +1,14 @@
+package auth
+
+import (
+	"context"
+
+	jwt "github.com/golang-jwt/jwt/v5"
+
+	"github.com/helixml/helix/api/pkg/types"
+)
+
+type Authenticator interface {
+	GetUserByID(ctx context.Context, userID string) (*types.User, error)
+	ValidateUserToken(ctx context.Context, token string) (*jwt.Token, error)
+}
