+++ api/pkg/config/config.go
@@ -13,7 +13,6 @@ type ServerConfig struct {
 	Providers          Providers
 	Tools              Tools
 	Keycloak           Keycloak
-	OIDC               OIDC
 	Notifications      Notifications
 	Janitor            Janitor
 	Stripe             Stripe
@@ -123,14 +122,6 @@ type Keycloak struct {
 	Realm               string `envconfig:"KEYCLOAK_REALM" default:"helix"`
 	Username            string `envconfig:"KEYCLOAK_USER"`
 	Password            string `envconfig:"KEYCLOAK_PASSWORD"`
-}
-
-type OIDC struct {
-	Enabled          bool   `envconfig:"OIDC_ENABLED" default:"false"` // If set, disables keycloak authentication
-	IssuerURL        string `envconfig:"OIDC_ISSUER"`
-	APIClientID      string `envconfig:"KEYCLOAK_CLIENT_ID" default:"api"`
-	ClientSecret     string `envconfig:"KEYCLOAK_CLIENT_SECRET"`
-	FrontEndClientID string `envconfig:"KEYCLOAK_FRONTEND_CLIENT_ID" default:"frontend"`
 }
 
 // Notifications is used for sending notifications to users when certain events happen
