+++ api/pkg/notification/notification.go
@@ -49,25 +49,25 @@ type Notifier interface {
 }
 
 type NotificationsProvider struct {
-	userRetriever auth.UserRetriever
+	authenticator auth.Authenticator
 
 	email *Email
 }
 
-func New(cfg *config.Notifications, userRetriever auth.UserRetriever) (Notifier, error) {
+func New(cfg *config.Notifications, authenticator auth.Authenticator) (Notifier, error) {
 	email, err := NewEmail(cfg)
 	if err != nil {
 		return nil, err
 	}
 
 	return &NotificationsProvider{
-		userRetriever: userRetriever,
+		authenticator: authenticator,
 		email:         email,
 	}, nil
 }
 
 func (n *NotificationsProvider) Notify(ctx context.Context, notification *Notification) error {
-	user, err := n.userRetriever.GetUserByID(ctx, notification.Session.Owner)
+	user, err := n.authenticator.GetUserByID(ctx, notification.Session.Owner)
 	if err != nil {
 		return fmt.Errorf("failed to get user '%s' details: %w", notification.Session.Owner, err)
 	}
