+++ api/pkg/server/auth_middleware.go
@@ -5,8 +5,11 @@ import (
 	"errors"
 	"fmt"
 	"net/http"
+	"strings"
 
+	jwt "github.com/golang-jwt/jwt/v5"
 	"github.com/helixml/helix/api/pkg/auth"
+	"github.com/helixml/helix/api/pkg/config"
 	"github.com/helixml/helix/api/pkg/store"
 	"github.com/helixml/helix/api/pkg/types"
 )
@@ -21,32 +24,107 @@ var (
 )
 
 var (
+	ErrNoAPIKeyFound           = errors.New("no API key found")
 	ErrNoUserIDFound           = errors.New("no user ID found")
 	ErrAppAPIKeyPathNotAllowed = errors.New("path not allowed for app API keys, use your personal account key from your /account page instead")
 )
 
-type authMiddleware struct {
-	oidcAuth auth.OIDCAuthenticator
-	runnerAuth auth.RunnerTokenAuthenticator
-	apikeyAuth auth.ApiKeyAuthenticator
+type authMiddlewareConfig struct {
+	adminUserIDs []string
+	adminUserSrc config.AdminSrcType
+	runnerToken  string
+}
 
-	store store.Store
+type authMiddleware struct {
+	authenticator auth.Authenticator
+	store         store.Store
+	cfg           authMiddlewareConfig
 }
 
 func newAuthMiddleware(
-	oidcAuth auth.OIDCAuthenticator,
-	runnerAuth auth.RunnerTokenAuthenticator,
-	apikeyAuth auth.ApiKeyAuthenticator,
-
+	authenticator auth.Authenticator,
 	store store.Store,
+	cfg authMiddlewareConfig,
 ) *authMiddleware {
 	return &authMiddleware{
-		oidcAuth: oidcAuth,
-		runnerAuth: runnerAuth,
-		apikeyAuth: apikeyAuth,
-
-		store: store,
+		authenticator: authenticator,
+		store:         store,
+		cfg:           cfg,
 	}
+}
+
+type tokenAcct struct {
+	jwt    *jwt.Token
+	userID string
+}
+
+type account struct {
+	userID string
+	token  *tokenAcct
+}
+
+type accountType string
+
+const (
+	accountTypeUser    accountType = "user"
+	accountTypeToken   accountType = "token"
+	accountTypeInvalid accountType = "invalid"
+)
+
+func (a *account) Type() accountType {
+	switch {
+	case a.userID != "":
+		return accountTypeUser
+	case a.token != nil:
+		return accountTypeToken
+	}
+	return accountTypeInvalid
+}
+
+func (auth *authMiddleware) isAdmin(acct account) bool {
+	if acct.Type() == accountTypeInvalid {
+		return false
+	}
+
+	switch auth.cfg.adminUserSrc {
+	case config.AdminSrcTypeEnv:
+		if acct.Type() == accountTypeUser {
+			return auth.isUserAdmin(acct.userID)
+		}
+		return auth.isUserAdmin(acct.token.userID)
+	case config.AdminSrcTypeJWT:
+		if acct.Type() != accountTypeToken {
+			return false
+		}
+		return auth.isTokenAdmin(acct.token.jwt)
+	}
+	return false
+}
+
+func (auth *authMiddleware) isUserAdmin(userID string) bool {
+	if userID == "" {
+		return false
+	}
+
+	for _, adminID := range auth.cfg.adminUserIDs {
+		// development mode everyone is an admin
+		if adminID == types.AdminAllUsers {
+			return true
+		}
+		if adminID == userID {
+			return true
+		}
+	}
+	return false
+}
+
+func (auth *authMiddleware) isTokenAdmin(token *jwt.Token) bool {
+	if token == nil {
+		return false
+	}
+	mc := token.Claims.(jwt.MapClaims)
+	isAdmin := mc["admin"].(bool)
+	return isAdmin
 }
 
 func (auth *authMiddleware) getUserFromToken(ctx context.Context, token string) (*types.User, error) {
@@ -54,28 +132,67 @@ func (auth *authMiddleware) getUserFromToken(ctx context.Context, token string) 
 		return nil, nil
 	}
 
-	result, err := auth.runnerAuth.ValidateAndReturnUser(ctx, token)
-	if err != nil {
-		return nil, fmt.Errorf("runner token auth caused an error: %s", err)
-	}
-	if result != nil {
-		return result, nil
-	}
-
-	result, err = auth.apikeyAuth.ValidateAndReturnUser(ctx, token)
-	if err != nil {
-		return nil, fmt.Errorf("api key auth caused an error: %s", err)
-	}
-	if result != nil {
-		return result, nil
+	if token == auth.cfg.runnerToken {
+		// if the api key is our runner token then we are in runner mode
+		return &types.User{
+			Token:     token,
+			TokenType: types.TokenTypeRunner,
+		}, nil
 	}
 
-	result, err = auth.oidcAuth.ValidateAndReturnUser(ctx, token)
-	if err != nil {
-		return nil, fmt.Errorf("oidc auth caused an error: %s", err)
+	if strings.HasPrefix(token, types.APIKeyPrefix) {
+		// we have an API key - we should load it from the database and construct our user that way
+		apiKey, err := auth.store.GetAPIKey(ctx, token)
+		if err != nil {
+			return nil, fmt.Errorf("error getting API key: %s", err.Error())
+		}
+		if apiKey == nil {
+			return nil, fmt.Errorf("error getting API key: %w", ErrNoAPIKeyFound)
+		}
+
+		user, err := auth.authenticator.GetUserByID(ctx, apiKey.Owner)
+		if err != nil {
+			return user, fmt.Errorf("error loading user from keycloak: %s", err.Error())
+		}
+
+		user.Token = token
+		user.TokenType = types.TokenTypeAPIKey
+		user.ID = apiKey.Owner
+		user.Type = apiKey.OwnerType
+		user.Admin = auth.isAdmin(account{userID: user.ID})
+		if apiKey.AppID != nil && apiKey.AppID.Valid {
+			user.AppID = apiKey.AppID.String
+		}
+
+		return user, nil
 	}
 
-	return result, nil
+	// otherwise we try to decode the token with keycloak
+	keycloakJWT, err := auth.authenticator.ValidateUserToken(ctx, token)
+	if err != nil {
+		return nil, fmt.Errorf("error validating keycloak token: %s", err.Error())
+	}
+	mc := keycloakJWT.Claims.(jwt.MapClaims)
+	keycloakUserID := mc["sub"].(string)
+
+	if keycloakUserID == "" {
+		return nil, fmt.Errorf("error getting keycloak user ID: %w", ErrNoUserIDFound)
+	}
+
+	user, err := auth.authenticator.GetUserByID(ctx, keycloakUserID)
+	if err != nil {
+		return user, fmt.Errorf("error loading user from keycloak: %s", err.Error())
+	}
+
+	user.Token = token
+	user.TokenType = types.TokenTypeKeycloak
+	user.ID = keycloakUserID
+	user.Type = types.OwnerTypeUser
+	user.Admin = auth.isAdmin(account{
+		token: &tokenAcct{jwt: keycloakJWT, userID: user.ID},
+	})
+
+	return user, nil
 }
 
 // this will extract the token from the request and then load the correct
