+++ api/pkg/server/server.go
@@ -78,12 +78,6 @@ type HelixAPIServer struct {
 	pingService       *version.PingService
 }
 
-type AuthConfig struct {
-	OIDCAuth auth.OIDCAuthenticator
-	RunnerAuth auth.RunnerTokenAuthenticator
-	ApiKeyAuth auth.ApiKeyAuthenticator
-}
-
 func NewServer(
 	cfg *config.ServerConfig,
 	store store.Store,
@@ -91,7 +85,7 @@ func NewServer(
 	gptScriptExecutor gptscript.Executor,
 	providerManager manager.ProviderManager,
 	inferenceServer *openai.InternalHelixServer,
-	authConfig *AuthConfig,
+	authenticator auth.Authenticator,
 	stripe *stripe.Stripe,
 	controller *controller.Controller,
 	janitor *janitor.Janitor,
@@ -124,10 +118,13 @@ func NewServer(
 		gptScriptExecutor: gptScriptExecutor,
 		inferenceServer:   inferenceServer,
 		authMiddleware: newAuthMiddleware(
-			authConfig.OIDCAuth,
-			authConfig.RunnerAuth,
-			authConfig.ApiKeyAuth,
+			authenticator,
 			store,
+			authMiddlewareConfig{
+				adminUserIDs: cfg.WebServer.AdminIDs,
+				adminUserSrc: cfg.WebServer.AdminSrc,
+				runnerToken:  cfg.WebServer.RunnerToken,
+			},
 		),
 		providerManager:  providerManager,
 		pubsub:           ps,
